# openapi/AiFramework.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "AiFramework.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/monitoring/jobs/dead-letters": {
      "get": {
        "tags": [
          "MonitoringDeadLetters"
        ],
        "summary": "The dead-letter queue.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DeadLetterPageResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeadLetterPageResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeadLetterPageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/monitoring/jobs/dead-letters/{messageId}/retry": {
      "post": {
        "tags": [
          "MonitoringDeadLetters"
        ],
        "summary": "Puts one dead-lettered message back in play.",
        "description": "The API marks the stored envelope replayable and stops there; the worker picks it up from\nthe shared PostgreSQL store on its own. The two hosts never talk, which is what makes this\npage possible without exposing the worker through the ingress. See ADR 0016.",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/monitoring/sign-ins": {
      "get": {
        "tags": [
          "MonitoringSignIns"
        ],
        "summary": "The sign-in audit trail, newest first.",
        "parameters": [
          {
            "name": "outcome",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SignInOutcome"
            }
          },
          {
            "name": "username",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SignInEventPageResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignInEventPageResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignInEventPageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeadLetterPageResponse": {
        "required": [
          "items",
          "totalCount",
          "page"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeadLetterResponse"
            }
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "DeadLetterResponse": {
        "required": [
          "id",
          "messageType",
          "exceptionType",
          "exceptionMessage",
          "sentAt",
          "replayable"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "messageType": {
            "type": "string"
          },
          "exceptionType": {
            "type": "string"
          },
          "exceptionMessage": {
            "type": "string"
          },
          "sentAt": {
            "type": "string",
            "format": "date-time"
          },
          "receivedAt": {
            "type": [
              "null",
              "string"
            ]
          },
          "replayable": {
            "type": "boolean",
            "description": "True once an operator has asked for it to be retried."
          }
        },
        "description": "A message that exhausted its retries, from Wolverine's own dead-letter queue."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SignInEventPageResponse": {
        "required": [
          "items",
          "totalCount",
          "page"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SignInEventResponse"
            }
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "SignInEventResponse": {
        "required": [
          "id",
          "at",
          "usernameAttempted",
          "outcome"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "at": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": [
              "null",
              "string"
            ],
            "description": "Null when the attempt named an account that does not exist.",
            "format": "uuid"
          },
          "usernameAttempted": {
            "type": "string",
            "description": "What was typed. The only record of an attempt against a name that does not exist."
          },
          "outcome": {
            "description": "The distinction the sign-in endpoint itself never reveals: it answers every failure\nidentically, because a differing answer is an account-enumeration oracle (ADR 0006). An\noperator investigating an attack needs the difference; an attacker must not have it.",
            "$ref": "#/components/schemas/SignInOutcome"
          },
          "ipAddress": {
            "type": [
              "null",
              "string"
            ]
          },
          "userAgent": {
            "type": [
              "null",
              "string"
            ]
          },
          "traceId": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "One attempt to authenticate."
      },
...
      "SignInOutcome": {
        "enum": [
          "Succeeded",
          "BadCredentials",
          "LockedOut",
          "UnknownUser",
          "SignedOutEverywhere"
        ],
        "description": "What happened on one attempt to authenticate."
      },
...
    }
  },
...
}